backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python
numpy
pandas
openpyxl
psycopg2-binary
python-multipart
websockets
python-dotenv
ffmpeg-python
watchfiles
aiofiles==23.2.1